Moved CSVUploader to core components folder
